Honor the semantics of set_attributes() and clear all previous attributes.
Since gtk_cell_layout_set_attributes() is a varargs function, we have to
reimplement it, and we missed this part the first time around.

Also change the parameter testing code to allow an empty key/val pair
list.  After mailing list discussion, it was deemed that this would be
useful for avoiding special cases in generic code that uses arbitrary
lists of attributes.

svn path=/trunk/; revision=1874