Make t/GtkCellRenderer.t more robust
The tests that check that all vfuncs of custom cell renderers are called
appropriately were relying on certain behavior of gtk+.  Apparently, that is
not appropriate on all platforms, as we have failure reports from armhf, mips
and mipsel.  Make the tests more robust by triggering all vfuncs on all custom
renderers manually.

https://bugzilla.gnome.org/show_bug.cgi?id=729453